# Inline nice function (doubles read performance)