--- a/Cargo.toml
+++ b/Cargo.toml
@@ -17,7 +17,7 @@ travis-ci = { repository = "Amanieu/thread_local-rs" }
 once_cell = "1.5.2"
 
 # This is actually a dev-dependency, see https://github.com/rust-lang/cargo/issues/1596
-criterion = { version = "0.3.3", optional = true }
+criterion = { version = "0.4.0", optional = true }
 
 [dev-dependencies]