Output from "cargo bench --features criterion":
Before: [2.6697 ns 2.6897 ns 2.7103 ns]
After: [1.3435 ns 1.3484 ns 1.3533 ns]

this can be improved further (but sadly this requires nightly):
[1.0103 ns 1.0150 ns 1.0213 ns]