@Amanieu 2022-11-12
Nice! What is the nightly-only change that improves performance even more?