@terrarier2111 2022-11-12
> 

It's using the thread_local attribute instead of using the thread_local macro, i have a working version of this improvement in the main branch of my fork.