@Amanieu 2022-11-12
A `nightly` feature flag would be good for this, yes.