@terrarier2111 2022-11-12
> A `nightly` feature flag would be good for this, yes.

sure, i'll do that but i'd propose to move this pr independently as i'll probably need a couple of days for the other thingy and its easier if this is already merged (also this improvement itself is big enough to warrant an own pr) also could you make a new  release once my second pr is merged?